Move tests, add core functionality for Wordsmith.
I had an idea for a 'one-shot only' character brewing for quite a while, and in this case, I thought there was enough merit to justify trying to implement this. After investigation of how temporary triggers were stored and used, I confirmed this was at least plausible, so here's the basic form of Wordsmith (name TBD) and the necessary format for the initial cards. There's some convoluted logic to make sure damage modifiers only apply to the initial damage.